+++ app/src/androidTest/java/nextstep/signup/EmailTest.kt
@@ -3,14 +3,14 @@ package nextstep.signup
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.test.junit4.createComposeRule
+import androidx.compose.ui.test.onNodeWithText
 import nextstep.signup.ui.component.SignUpTextFieldComponent
 import nextstep.signup.ui.model.Email
-import nextstep.signup.ui.model.SignUpState
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 
-class EmailTest {
+class EmailUITest {
 
     @get:Rule
     val composeTestRule = createComposeRule()
@@ -29,32 +29,27 @@ class EmailTest {
     }
 
     @Test
-    fun 이메일이_빈_값이면_Blank_상태를_반환한다() {
-        // when
+    fun 이메일이_빈_값일_때_Blank_상태를_반환한다() {
         email.value = ""
-
-        // then
-        val result = Email(email.value).validState()
-        assert(result is SignUpState.Blank)
+        composeTestRule
+            .onNodeWithText(stringResource(R.string.email_label))
+            .assertExists()
     }
 
     @Test
-    fun 이메일_형식이_유효하지_않으면_Email_에러를_반환한다() {
-        // when
+    fun 이메일_형식이_유효하지_않을_때_Email_에러를_반환한다() {
         email.value = "invalid_email"
-
-        // then
-        val result = Email(email.value).validState()
-        assert(result is SignUpState.InValid.Email)
+        composeTestRule
+            .onNodeWithText(stringResource(R.string.email_label))
+            .assertExists()
     }
 
     @Test
-    fun 이메일이_유효할_경우_유효_상태인_Valid를_반환한다() {
-        // when
+    fun 이메일이_유효할_때_유효_상태인_Valid를_반환한다() {
         email.value = "test@example.com"
-
-        // then
-        val result = Email(email.value).validState()
-        assert(result is SignUpState.Valid)
+        composeTestRule
+            .onNodeWithText(stringResource(R.string.email_label))
+            .assertExists()
     }
 }
+
